FINALLY WORKING USER AUTHORIZATION

+++ backend/src/main/java/com/expense_tracker/exceptions/categories/CategoryNotFoundException.java
@@ -1,5 +1,5 @@
 package com.expense_tracker.exceptions.categories;
 
 public class CategoryNotFoundException extends RuntimeException {
-    public CategoryNotFoundException(int id) { super("Could not find category with id: " + id); }
+    public CategoryNotFoundException(Long id) { super("Could not find category with id: " + id); }
 }
