+++ backend/src/main/java/com/expense_tracker/exceptions/expenses/ExpenseNotFoundException.java
@@ -1,7 +1,7 @@
 package com.expense_tracker.exceptions.expenses;
 
 public class ExpenseNotFoundException extends RuntimeException{
-    public ExpenseNotFoundException(int id) {
+    public ExpenseNotFoundException(Long id) {
         super("Could not find expense with id: " + id);
     }
 }
