+++ backend/src/main/java/com/expense_tracker/exceptions/users/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package com.expense_tracker.exceptions.users;
 
 public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(Long id) {
         super("Could not find user with id: " + id);
     }
 }
