+++ backend/src/main/java/com/expense_tracker/repositories/ExpensesRepository.java
@@ -1,12 +1,12 @@
 package com.expense_tracker.repositories;
 
-import com.expense_tracker.entities.Expense;
+import com.expense_tracker.models.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpensesRepository extends JpaRepository<Expense, Integer> {
+public interface ExpensesRepository extends JpaRepository<Expense, Long> {
 
-    Iterable<Expense> findByUserId(Integer userId);
+    Iterable<Expense> findByUserId(Long userId);
 
 }
